+++ daemon/libnetwork/network_unix.go
@@ -3,7 +3,6 @@
 package libnetwork
 
 import (
-	"context"
 	"fmt"
 	"strconv"
 	"time"
@@ -19,16 +18,6 @@ type platformNetwork struct{} //nolint:nolintlint,unused // only populated on wi
 // Stub implementations for DNS related functions
 
 func (n *Network) startResolver() {
-}
-
-func addEpToResolver(
-	ctx context.Context,
-	netName, epName string,
-	config *containerConfig,
-	epIface *EndpointInterface,
-	resolvers []*Resolver,
-) error {
-	return nil
 }
 
 func deleteEpFromResolver(epName string, epIface *EndpointInterface, resolvers []*Resolver) error {
