+++ daemon/libnetwork/sandbox_windows.go
@@ -4,6 +4,7 @@ import (
 	"context"
 
 	"github.com/moby/moby/v2/daemon/libnetwork/osl"
+	"github.com/moby/moby/v2/errdefs"
 )
 
 func releaseOSSboxResources(*osl.Namespace, *Endpoint) {}
@@ -37,8 +38,11 @@ func (sb *Sandbox) canPopulateNetworkResources() bool {
 	return true
 }
 
-func (sb *Sandbox) populateNetworkResources(context.Context, *Endpoint) error {
-	// not implemented on Windows (Sandbox.osSbox is always nil)
+func (sb *Sandbox) populateNetworkResourcesOS(ctx context.Context, ep *Endpoint) error {
+	n := ep.getNetwork()
+	if err := addEpToResolver(ctx, n.Name(), ep.Name(), &sb.config, ep.iface, n.Resolvers()); err != nil {
+		return errdefs.System(err)
+	}
 	return nil
 }
